Use ints for invlets (#31468)
Invlets were previously longs.  They are just characters or keys, so
ints are sufficient.

+++ src/character.h
@@ -366,7 +366,7 @@ class Character : public Creature, public visitable<Character>
         /** Returns true if player has a trait with a flag */
         bool has_trait_flag( const std::string &flag ) const;
         /** Returns the trait id with the given invlet, or an empty string if no trait has that invlet */
-        trait_id trait_by_invlet( long ch ) const;
+        trait_id trait_by_invlet( int ch ) const;
 
         /** Toggles a trait on the player and in their mutation list */
         void toggle_trait( const trait_id &flag );
